leggg testing av utbetalingsperioder i StønadsstatistikkServiceTest.kt
